Split TypeScriptStubs tests
